Move all the construct-related code into one file
