+++ src/vg.hpp
@@ -161,6 +161,10 @@ public:
     // construct from sets of nodes and edges (e.g. subgraph of another graph)
     VG(set<Node*>& nodes, set<Edge*>& edges);
 
+
+    // Now all the stuff for constructing from VCF.
+    // TODO: refactor out
+
     // construct from VCF
     VG(vcflib::VariantCallFile& variantCallFile,
        FastaReference& reference,
@@ -190,6 +194,30 @@ public:
                        int start_pos,
                        int stop_pos,
                        int max_node_size);
+                       
+    // Takes in a VCF file
+    // and returns a map [node] = vcflib::variant
+    // Unfortunately this is specific to a given graph
+    // and VCF.
+    //
+    // It will need to throw warnings if the node or variant
+    // is not in the graph.
+    //
+    // This is useful for VCF masking:
+    // if map.find(node) then mask variant
+    //
+    // It's also useful for calling known variants
+    // for m in alignment.mappings:
+    //    node = m.Pos.nodeID
+    //    if node in node_to_vcf:
+    //        return (alignment supports variant)
+    //
+    // It would be nice if this also supported edges (e.g.
+    // for inversions/transversions/breakpoints?)
+    // map<edge_id, variant> or map<pair<NodeID, NodeID>, variant>
+    map<id_t, vcflib::Variant> get_node_id_to_variant(vcflib::VariantCallFile vfile);
+                       
+                       
     // chops up the nodes
     void dice_nodes(int max_node_size);
     // does the reverse --- combines nodes by removing edges where doing so has no effect on the graph labels
@@ -601,29 +629,6 @@ public:
     // (e.g. from a vcf) and returns the node id which
     // contains that position.
     id_t get_node_at_nucleotide(string pathname, int nuc);
-
-    // Takes in a VCF file
-    // and returns a map [node] = vcflib::variant
-    // Unfortunately this is specific to a given graph
-    // and VCF.
-    //
-    // It will need to throw warnings if the node or variant
-    // is not in the graph.
-    //
-    // This is useful for VCF masking:
-    // if map.find(node) then mask variant
-    //
-    // It's also useful for calling known variants
-    // for m in alignment.mappings:
-    //    node = m.Pos.nodeID
-    //    if node in node_to_vcf:
-    //        return (alignment supports variant)
-    //
-    // It would be nice if this also supported edges (e.g.
-    // for inversions/transversions/breakpoints?)
-    // map<edge_id, variant> or map<pair<NodeID, NodeID>, variant>
-    map<id_t, vcflib::Variant> get_node_id_to_variant(vcflib::VariantCallFile vfile);
-
 
     // edges
     // If the given edge cannot be created, returns null.
